Added ConvNet back to Recurrent net
ConvNets can be combined with current nets again. Also, wrotesuperconv6
that does high speed temporal gradient for conv net.

+++ nn_core/cuda/cudaBuild_and_Test.m
@@ -1,8 +1,12 @@
 clear
 addpath('../');
 gDev = gpuDevice(1);
-system('rm *.ptx','-echo');
-PRECISION = 'double';
+if (isunix)
+    system('rm *.ptx','-echo');
+else
+    system('del *.ptx','-echo');
+end
+PRECISION = 'single';
 Nbench = 1;
 
 disp('Building, testing and benchmarking NVIDIA/CUDA accelerated convolutional net kernels.');
@@ -11,7 +15,7 @@ disp('  -> NVIDIA Cuda Toolkit 7.0 and either GCC or Visual Studio 2013 compiler
 %%%%%%%%%%%%%%%%%%%%%%%%%%% SUPERCONV5 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %%{
 disp('1. Compiling superconv5 (used in cnnUnconvolve)');
-[status,result] = system('nvcc --compiler-bindir /usr/bin/gcc-4.9 -ptx superconv5.cu','-echo');
+[status,result] = system('nvcc -ptx superconv5.cu','-echo');
 if status == 0
     disp('Running error test');
     Ni = int32(16); % number of images
@@ -38,7 +42,7 @@ if status == 0
     err = (out_conv - out).^2;
     err = sum(err(:));
     disp(['Numerical error is ' num2str(err) '.']);
-    if err > 1e-16
+    if err > 1e-6
         error('superconv5 failed numerical accuracy check!');
     end
     
@@ -80,7 +84,7 @@ disp('-----------------'); disp(' ');
 %%%%%%%%%%%%%%%%%%%%%%%%%%% SUPERCONV4 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %%{
 disp('2. Compiling superconv4 (used in cnnConvGrad)');
-[status,result] = system('nvcc --compiler-bindir /usr/bin/gcc-4.9 -ptx superconv4.cu','-echo');
+[status,result] = system('nvcc -ptx superconv4.cu','-echo');
 if status == 0
     disp('Running error test');
     Ni = int32(16); % number of images
@@ -104,7 +108,7 @@ if status == 0
     err = (out_conv - out).^2;
     err = sum(err(:));
     disp(['Numerical error is ' num2str(err) '.']);
-    if err > 1e-16
+    if err > 1e-6
         error('superconv4 failed numerical accuracy check!');
     end
     
@@ -141,10 +145,86 @@ disp('-----------------'); disp(' ');
 %}
 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 
+%%%%%%%%%%%%%%%%%%%%%%%%%%% SUPERCONV6 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+%%{
+disp('3. Compiling superconv6 (used in cnnConvGradTemporal)');
+[status,result] = system('nvcc -ptx superconv6.cu','-echo');
+if status == 0
+    disp('Running error test');
+    Ni = int32(9); % number of images
+    Nt = int32(10); % time slices
+    %% NOTE!! NOT HANDLING EVEN Nt! (must be mod arithmetic)
+    Nd = int32([9 6 5 Ni Nt]); % dimension of kernel
+    Na = int32([28 28 7 Ni Nt]); % dimension of image
+   
+    A = gpuArray.randn(Na, PRECISION);
+    d = gpuArray.randn(Nd, PRECISION);
+    
+    % test output for accuracy
+    p = [];
+    [out, p] = superconv6(A, d, p);
+    out = sum(sum(out,6),5);
+    
+    out_conv_t = gpuArray.zeros(size(out), PRECISION);
+    out_conv = out_conv_t;
+    for t = 1:Nt
+        for j=1:Na(3) % loop over input maps
+            for i=1:Nd(3) % loop over output maps
+                out_conv_t(:,:,j,i) = convn((A(:,:,j,:,t)), flipall3(d(:,:,i,:,t), 4), 'valid');
+            end
+        end 
+        out_conv = out_conv + out_conv_t;
+    end
+    
+    err = (out_conv - out).^2;
+    err = sum(err(:));
+    disp(['Numerical error is ' num2str(err) '.']);
+    if err > 1e-5
+        error('superconv6 failed numerical accuracy check!');
+    end
+    
+    % benchmark
+    disp('Running benchmarks');
+    tic 
+    for i=coder.unroll(1:Nbench, true)
+        [out, p] = superconv6(A, d, p);
+        out = sum(sum(out, 6),5);
+    end
+    wait(gDev);
+    tgpu = toc;
+    disp([num2str(tgpu/Nbench*1e3) ' ms for kernel']);
+    
+    %d = gather(flipall3(d,3));
+    %A = gather(A);
+    tic
+    for i=coder.unroll(1:Nbench, true)
+        out_conv_t = gpuArray.zeros(size(out), PRECISION);
+        out_conv = out_conv_t;
+        for t = 1:Nt
+            for j=1:Na(3) % loop over input maps
+                for i=1:Nd(3) % loop over output maps
+                    out_conv_t(:,:,j,i) = convn((A(:,:,j,:,t)), flipall3(d(:,:,i,:,t), 4), 'valid');
+                end
+            end 
+            out_conv = out_conv + out_conv_t;
+        end
+    end
+    wait(gDev);
+    tcpu = toc;
+    disp([num2str(tcpu/Nbench*1e3) ' ms for convn (matlab)']);
+    disp(['Speed up is: ' num2str(tcpu/tgpu) 'X for superconv6']);
+else
+    error('Error compiling superconv6!');
+end
+clear out_conv;
+disp('-----------------'); disp(' ');
+%}
+%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
+
 %%%%%%%%%%%%%%%%%%%%%%%%%%% SUPERCONV3 %%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%%
 %%{
-disp('3. Compiling superconv3 (used in cnnConvolve)');
-[status,result] = system('nvcc --compiler-bindir /usr/bin/gcc-4.9 -ptx superconv3.cu','-echo');
+disp('4. Compiling superconv3 (used in cnnConvolve)');
+[status,result] = system('nvcc -ptx superconv3.cu','-echo');
 if status == 0
     disp('Running error test');
     Nk = int32([6 5 10]); % dimension of kernel
@@ -162,7 +242,7 @@ if status == 0
     err = (out_conv - out).^2;
     err = sum(err(:));
     disp(['Numerical error is ' num2str(err) '.']);
-    if err > 1e-16
+    if err > 1e-6
         error('superconv3 failed numerical accuracy check!');
     end
